pyproject.toml
[tool.poetry]
name = "geekhunter-backend"
version = "0.1.0"
description = ""
authors = ["'Yaroslav <'arhdomain@gmail.com'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
beautifulsoup4 = "^4.12.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.7"
ruff = "^0.0.290"
xmltodict = "^0.13.0"
pre-commit = "^3.4.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
tqdm = "^4.66.1"
pydantic = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",
    "F",
    "W",
    "N",
    "D",
    "UP",
    "ANN",
    "BLE",
    "B",
    "C4",
    "DJ",
    "ICN",
    "PYI",
    "RET",
    "SIM",
    "PL",
]

exclude = [
    "migrations",
    "__init__.py",
    "asgi.py",
    "wsgi.py",
    "manage.py",
    "settings.py",
    "apps.py",
]

ignore = [
    "D100",
    "D107",
    "D106",
    "D400",
    "D415",
    "ANN101",
    "DJ001",
    "D105",
    "D203",
    "D213",
    "PLR0912",
    "PLR0915",
]

line-length = 79
